+++ mod_form.php
@@ -46,14 +46,7 @@ class mod_kialo_mod_form extends moodleform_mod {
      * @return string
      */
     private function get_deployment_id(): string {
-        global $COURSE;
-        global $USER;
-
-        // Since the deployment id corresponds to an activity id, but the activity hasn't been created yet,
-        // when the deep linking happens, we need to use a different deployment id.
-        $deploymentid = uniqid($COURSE->id . $USER->id, true);
-
-        return $deploymentid;
+        return "1";
     }
 
     /**
@@ -102,6 +95,11 @@ class mod_kialo_mod_form extends moodleform_mod {
         $mform->setType("discussion_title", PARAM_TEXT);
         $mform->addRule('discussion_title', null, 'required', null, 'client');
 
+        // Hidden copy of discussion URL filled by deeplinking. Form can only be submitted if this matches the field above,
+        // which means the user successfully linked the discussion via the deeplinking button.
+        $mform->addElement("hidden", "discussion_url", "");
+        $mform->setType("discussion_url", PARAM_RAW);
+
         // Deployment ID, filled when selecting the discussion.
         $deploymentid = $this->get_deployment_id();
         $mform->addElement("hidden", "deployment_id", $deploymentid);
@@ -126,6 +124,7 @@ class mod_kialo_mod_form extends moodleform_mod {
                   if (event.data.type !== 'kialo_discussion_selected') return;
 
                   // Fill in the deep-linked details.
+                  document.querySelector('input[name=discussion_url]').value = event.data.discussionurl;
                   document.querySelector('input[name=deployment_id]').value = event.data.deploymentid;
                   document.querySelector('input[name=discussion_title]').value = event.data.discussiontitle;