+++ tests/generator/lib.php
@@ -43,6 +43,9 @@ class mod_kialo_generator extends testing_module_generator {
         if (!isset($record->discussion_title)) {
             $record->discussion_title = "Test discussion";
         }
+        if (!isset($record->discussion_url)) {
+            $record->discussion_url = "https://www.kialo-edu.com/42";
+        }
 
         return parent::create_instance($record, $options);
     }